project reorganization: avoid nested projects
- keep flat tree
- target separation
- .so file included into jar file
- compile executable jar per example (deps and no-deps included)
- readme files update
