+++ examples/ByteBufferExample/src/main/java/ByteBufferExample.java
