+++ examples/MixedTypesExample/src/main/java/MixedTypesExample.java
