+++ examples/PicturesExample/src/main/java/PicturesExample.java
