+++ examples/StringExample/src/main/java/StringExample.java
