+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/BuilderException.java
