+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/DatabaseException.java
