+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/KeyCallback.java
