+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/KeyValueCallback.java
