+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/NotFoundException.java
