+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/NotSupportedException.java
