+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/OutOfMemoryException.java
