+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/StoppedByCallbackException.java
