+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/TransactionScopeException.java
