+++ pmemkv-binding/src/main/java/io/pmem/pmemkv/ValueCallback.java
