+++ pmemkv-binding/src/main/resources/version.properties
