--- src/main/pom.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.pmem</groupId>
-        <artifactId>pmemkv-java</artifactId>
-        <version>1.0.0</version>
-        <relativePath>../..</relativePath>
-    </parent>
-    <groupId>io.pmem</groupId>
-    <artifactId>pmemkv</artifactId>
-    <name>pmemkv</name>
-    <packaging>jar</packaging>
-    <description>Java code and main package for pmemkv Java bindings</description>
-
-    <build>
-        <sourceDirectory>java</sourceDirectory>
-        <testSourceDirectory>${basedir}/../test</testSourceDirectory>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                  <showWarnings>true</showWarnings>
-                  <failOnWarning>true</failOnWarning>
-                  <compilerArgs>
-                    <arg>-Xlint:all</arg>
-                  </compilerArgs>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.22.1</version>
-                <configuration>
-                    <useSystemClassLoader>false</useSystemClassLoader>
-                    <argLine>-Djava.library.path=${project.build.directory}/classes -Dfile.encoding=UTF-8</argLine>
-                </configuration>
-            </plugin>
-            <plugin>
-                <artifactId>maven-resources-plugin</artifactId>
-                <version>3.1.0</version>
-                <executions>
-                  <execution>
-                    <id>copy-resources</id>
-                    <phase>compile</phase>
-                    <goals>
-                      <goal>copy-resources</goal>
-                    </goals>
-                    <configuration>
-                      <outputDirectory>${project.build.directory}/classes</outputDirectory>
-                      <resources>
-                        <resource>
-                          <directory>${basedir}/cpp/target</directory>
-                          <includes>
-                            <include>*.so*</include>
-                          </includes>
-                          <filtering>false</filtering>
-                        </resource>
-                      </resources>
-                    </configuration>
-                  </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
-                <version>3.0.0</version>
-            </plugin>
-        </plugins>
-        <resources>
-            <resource>
-                <directory>${basedir}/resources</directory>
-                <includes>
-                    <include>version.properties</include>
-                </includes>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-    </build>
-
-    <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>[4.12,)</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>[1.10,)</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-</project>
