+++ utils/docker/run-build.sh
@@ -13,10 +13,10 @@ source `dirname $0`/prepare-for-build.sh
 
 function run_example() {
 	example_name=$1
-	jar_path=../src/main/target/pmemkv-1.0.0.jar
-
-	javac -g -cp ${jar_path} ${example_name}.java
-	java -ea -Xms1G -cp .:${jar_path} ${example_name}
+	jar_path=../pmemkv-binding/target/pmemkv-1.0.0.jar
+# Find path to a jar with specific example name
+  example_path=`find .. | grep -P '\b(?!pmemkv)\b([a-zA-Z]+)\-([0-9.]+)\.jar' | grep ${example_name}`
+	java -ea -Xms1G -cp ${jar_path}:${example_path} ${example_name}
 }
 
 # install pmemkv
@@ -44,8 +44,6 @@ echo
 echo "###########################################################"
 echo "### Verifying building and execution of examples"
 echo "###########################################################"
-# set path to libpmemkv-jni.so
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$WORKDIR/src/main/cpp/target
 cd examples
 run_example StringExample
 run_example ByteBufferExample
